[Blacklist] Refactor the blacklist in TypeScript
