+++ app/javascript/entrypoints/v_posts_show.ts
@@ -5,6 +5,7 @@ declare const E621: E621Type;
 
 import "@/pages/comments/comments";
 import "@/pages/posts/show/AddToPoolDialog";
+import "@/pages/posts/show/BlacklistRevealOnClick";
 import "@/pages/posts/show/MobileTabs";
 import "@/pages/posts/show/mod_queue";
 import Note from "@/pages/posts/show/notes";
