+++ app/javascript/src/js/core/DeferredPostLoader.ts
@@ -33,9 +33,9 @@ export default class DeferredPostLoader {
     // At this point, no thumbnails are rendered.
     // However, this lets us populate the Blacklist and render thumbnails with the correct
     // visibility on the first try, instead of rendering them and then updating their visibility.
-    E621.Blacklist.load_deferred_posts(processed);
-    E621.Blacklist.update_visibility();
-    E621.Blacklist.update_styles();
+    E621.Blacklist.loadDeferredPosts(processed);
+    E621.Blacklist.updatePostVisibility();
+    E621.Blacklist.updateThumbnailStyles();
 
     this.Logger.log("Loaded post data", "\n ⤷ Processed: " + processed.length + " / " + Object.keys(postsData).length, "\n ⤷ In cache: " + PostCache.stats().cachedPosts);
