+++ app/javascript/src/js/pages/posts/show/BlacklistRevealOnClick.ts
@@ -1,0 +1,10 @@
+$(() => {
+  const container = $("#image-container")
+    .off("click.blacklist")
+    .on("click.blacklist", () => {
+      if (!container.hasClass("blacklisted")) return;
+      container.removeClass("blacklisted");
+
+      $("#note-container").css("visibility", "visible");
+    });
+});
