+++ app/views/appeals/types/_flag.html.erb
@@ -12,7 +12,7 @@
   <div class="appeal-display-section appeal-display-flag-post">
     <%= render(PostThumbnailComponent.new(post: @flag.post, stats: true, show_deleted: true)) %>
   </div>
-  <%= render "posts/partials/common/inline_blacklist" %>
+  <%= render partial: "application/blacklist_inline" %>
 <% else %>
   <div class="appeal-display-flag">
     Flag #<%= @appeal.disp_id %> has been destroyed.
