+++ app/views/application/_blacklist.html.erb
@@ -1,7 +1,11 @@
-<section
-  class="blacklist-ui blacklist-inline"
+<%# locals: (post_id: nil) %>
+
+<section class="blacklist-ui staged"
   filters="0"
   collapsed="true"
+  <% if post_id.present? %>
+    post="<%= post_id %>"
+  <% end %>
 >
   <div class="blacklist-header">Blacklisted <span class="blacklisted-count"></span></div>
   <ul class="blacklist-filters"></ul>
