+++ app/views/blips/index.html.erb
@@ -4,7 +4,7 @@
     <div id="blip-list">
       <%= render "search" %>
 
-      <%= render "posts/partials/common/inline_blacklist" %>
+      <%= render partial: "application/blacklist_inline" %>
       <% if @blips.empty? %>
         <h4>No blips</h4>
       <% else %>
