+++ app/views/moderator/dashboards/show.html.erb
@@ -3,7 +3,7 @@
     <h1>Moderator Dashboard</h1>
 
     <%= render "search" %>
-    <%= render "posts/partials/common/inline_blacklist" %>
+    <%= render partial: "application/blacklist_inline" %>
 
     <div class="activity-container">
       <div>
