+++ app/views/pools/gallery.html.erb
@@ -2,7 +2,7 @@
   <div id="a-gallery">
     <%= render partial: "search", locals: { path: gallery_pools_path } %>
 
-    <%= render "posts/partials/common/inline_blacklist" %>
+    <%= render partial: "application/blacklist_inline" %>
 
     <h2>Pools</h2>
