+++ app/views/pools/show.html.erb
@@ -9,7 +9,7 @@
       <%= format_text(@pool.description) %>
     </div>
 
-    <%= render "posts/partials/common/inline_blacklist" %>
+    <%= render partial: "application/blacklist_inline" %>
 
     <div id="posts">
       <section class="posts-container">
