+++ app/views/popular/index.html.erb
@@ -4,7 +4,7 @@
 
     <%= @post_set.presenter.nav_links(self) %>
 
-    <%= render "posts/partials/common/inline_blacklist" %>
+    <%= render partial: "application/blacklist_inline" %>
 
     <div id="posts">
       <section class="posts-container">
