+++ app/views/post_flags/index.html.erb
@@ -2,7 +2,7 @@
   <div id="a-index">
     <h1>Flags</h1>
     <%= render "search" %>
-    <%= render "posts/partials/common/inline_blacklist" %>
+    <%= render partial: "application/blacklist_inline" %>
     <table class="striped">
       <thead>
         <tr>
