+++ app/views/post_replacements/index.html.erb
@@ -3,7 +3,7 @@
     <h1>Post Replacements</h1>
 
     <%= render "search" %>
-    <%= render "posts/partials/common/inline_blacklist" %>
+    <%= render partial: "application/blacklist_inline" %>
 
     <table class="striped autofit" style="max-width:100%">
       <thead>
