+++ app/views/uploads/index.html.erb
@@ -1,7 +1,7 @@
 <div id="c-uploads">
   <div id="a-index">
     <%= render "uploads/search" %>
-    <%= render "posts/partials/common/inline_blacklist" %>
+    <%= render partial: "application/blacklist_inline" %>
 
     <table class="striped autofit">
       <thead>
