+++ app/views/wiki_pages/_sidebar.html.erb
@@ -1,4 +1,4 @@
 <aside id="sidebar">
-  <%= render "posts/partials/index/blacklist" %>
+  <%= render partial: "application/blacklist" %>
   <%= render "wiki_pages/recent_changes" %>
 </aside>
